Test-case: check-audio-equalizer: Updates for IPC4 systems
This patch updates the test check-audio-equalizer.sh to work
with IPC4 topologies with different naming for controls. The
control name no more includes the component name, so the
search for controls is changed. A tool topo_effect_kcontrols.py
is used to find bytes controls for given component.

The blobs to test with sof-ctl are updated from SOF git
repository. The ipc3 or ipc4 directory is chosen based on
runtime check of DUT ipc version.

The sof-ctl usage is changed. The control is now referenced
with the text control name instead of numid.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>